app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sqlite-utils==3.36
selenium==4.15.2
beautifulsoup4==4.12.2
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
fuzzywuzzy==0.18.0
python-levenshtein==0.21.1
aiofiles==23.2.1
websockets==12.0
